+++ packages/affiliate/src/pages/multiLevel/affiliate/multiLevelReferral/multiLevelReferral.widget.tsx
@@ -1,0 +1,7 @@
+import { useMultiLevelReferralScript } from "./multiLevelReferral.script";
+import { MultiLevelReferral } from "./multiLevelReferral.ui";
+
+export const MultiLevelReferralWidget = () => {
+  const state = useMultiLevelReferralScript();
+  return <MultiLevelReferral {...state} />;
+};