+++ packages/affiliate/src/pages/multiLevel/affiliate/referrerTable/refereesTable/index.ts
@@ -1,0 +1,1 @@
+export { RefereesTableWidget as RefereesTable } from "./refereesTable.widget";